add Jyx2LeaveNode
