+++ jyx2/Assets/Scripts/EventsGraph/Nodes/Jyx2LeaveNode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 087b371d20ec9b74eab2562323e1983a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
